+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and as of version 1.0.0, follows semantic versioning.
 
 ## [Unreleased]
   * One more column for the All Samples table: Coverage for single lane TAR
+  * Bcl2barcode view has been greatly simplified by using the `bcl2barcodecaller` gsiqcetl cache
 
 ## [230529-1503] - 2023-05-29
   * Yet more columns for the All Samples tables